dapp images refactoring, dapp description component
