+++ src/components/dapp-staking/register/DescriptionTab.vue
